+++ gradle-plugin-api/src/test/kotlin/io/spine/tools/gradle/lib/LibraryPluginSpec.kt
@@ -28,9 +28,9 @@ package io.spine.tools.gradle.lib
 
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
-import io.spine.tools.gradle.root.plugin.RootPlugin
-import io.spine.tools.gradle.root.plugin.hasRootExtension
-import io.spine.tools.gradle.root.plugin.rootExtension
+import io.spine.tools.gradle.root.RootPlugin
+import io.spine.tools.gradle.root.hasRootExtension
+import io.spine.tools.gradle.root.rootExtension
 import org.gradle.api.Project
 import org.gradle.api.UnknownDomainObjectException
 import org.gradle.testfixtures.ProjectBuilder
