+++ gradle-root-plugin/src/main/kotlin/io/spine/tools/gradle/root/SettingsExtension.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.gradle.root.plugin
+package io.spine.tools.gradle.root
 
 import org.gradle.api.plugins.ExtensionAware
 import org.gradle.api.provider.Property
