+++ gradle-root-plugin/src/main/kotlin/io/spine/tools/gradle/root/SettingsExts.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.gradle.root.plugin
+package io.spine.tools.gradle.root
 
 import org.gradle.api.initialization.Settings
 import org.gradle.kotlin.dsl.findByType
